Moved sqlite dll to database folder
Store all future database related files in the Database folder
