+++ ChildcareApplication/ChildcareApplication/AdminTools/AdminEditChildInfo.xaml
@@ -1,7 +1,7 @@
-﻿<Window x:Class="ChildCareApp.win_AdminEditChildInfo"
+﻿<Window x:Class="AdminTools.win_AdminEditChildInfo"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:local="clr-namespace:ChildCareApp"
+        xmlns:local="clr-namespace:AdminTools"
         Title="win_AdminEditChildInfo" Height="465" Width="656">
     <Grid>
         <Canvas x:Name="cnv_ChildIcon" HorizontalAlignment="Left" Height="152" Margin="23,41,0,0" VerticalAlignment="Top" Width="158"/>
