+++ ChildcareApplication/ChildcareApplication/AdminTools/ParentInfoDB.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace ChildCareApp {
+namespace AdminTools {
     class ParentInfoDB {
         private MySqlConnection connection;
