+++ ChildcareApplication/ChildcareApplication/AdminTools/win_EditEvents.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="ChildCareApp.win_EditEvents"
+﻿<Window x:Class="AdminTools.EditEvents"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="win_EditEvents" Height="331.5" Width="720.5">
