+++ ChildcareApplication/ChildcareApplication/AdminTools/win_EventModificationWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="ChildCareApp.win_EventModificationWindow"
+﻿<Window x:Class="AdminTools.win_EventModificationWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="win_EventModificationWindow" Height="309" Width="375">
