+++ ChildcareApplication/ChildcareApplication/AdminTools/win_Link_DeLinkChild.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="ChildCareApp.win_Link_DeLinkChild"
+﻿<Window x:Class="AdminTools.win_Link_DeLinkChild"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="win_Link_DeLinkChild" Height="300" Width="435">
