+++ ChildcareApplication/ChildcareApplication/AdminTools/win_Link_DeLinkChild.xaml.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using System.Data;
-namespace ChildCareApp
+namespace AdminTools
 {
     /// <summary>
     /// Interaction logic for win_Link_DeLinkChild.xaml
