+++ ChildcareApplication/ChildcareApplication/AdminTools/win_ParentLogin.xaml.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace ChildCareApp
+namespace AdminTools
 {
     /// <summary>
     /// Interaction logic for win_ParentLogin.xaml
